requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-braces==1.13.0
django-cors-headers==3.1.1
django-sslserver==0.21
djangorestframework==3.10.3
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
oauthlib==3.1.0
isort==4.3.19
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.3.4
matplotlib==3.1.1
mccabe==0.6.1
numpy==1.17.0
psycopg2==2.8.4
PyJWT==1.7.1
python3-openid==3.1.0
pylint==2.3.1
pyparsing==2.4.2
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.25.3
virtualenv==16.7.2
wrapt==1.11.1
whitenoise==4.1.4
